Entity framework and player movement
Added an entity superclass which should be a parent of all entities - player, enemy, bullet etc
Added player, weapon, bullet classes
Added input, logic and rendering in the update loop for entities
Player class has movement with a following camera.
Player movement is normalized so the player moves at the same speed diagonally as horizontally and vertically.
Added the ability to detect clicks, not currently in use.
Added a player texture and a texture to use when you do not have a texture.