Hit particles and layers for rendering entities
Previously entities were rendered in the order they were added. Now, there is a field in the Entity class, layer, which determines the order entities are rendered in. Entities with a lower layer value are rendered first, or underneath.
Added blood particles that spray out when zombies and the player are damaged.