Bug fixes, landmines, equipment shows pickup text
Fixed bug where getting to tier 4 would cause the game to crash because the tier index went out of bounds for calculating the enemy cap because I was adding instead of subtracting one.
Landmines - they explode when you or a zombie step on them, doing a lot of damage in an area.
Equipment now shows text saying what it does when you pick it up